[VAST] `ExternPackageType` & tighten space invariants.
Adds an `ExternPackageType` to refer to types defined in other packages (that
we do not know the underlying bit width for or have a definition to refer to in
the current `VerilogFile`).

In the process found it was hard to reason about the invariants of `Emit()`
giving back associated spaces for types and `EmitWithIdentifier()` and similar
-- reworked that so all Emit functionality gives back a value with no
whitespace to strip so the caller knows what to expect. Previously it was
confusing/confused because DataKind could be empty string and DataType could
also be empty string in various cases. Case in point this fixed an error in the
test where there were two spaces after the `output` keyword.

Also adds a bunch of tests for constructs we're about to use in stitching just
to ensure they were all possible.