[DSLX] Fix handling of `xN`, e.g. in IR converter.
`xN` was primarily created so we could write polymorphic metaprogramming on
sign like `sizeof` in the standard library. Since `xN` is not fuzzed beyond
that simple use there were opportunities for mishandling it in IR conversion,
and the DSLX language tests I believe don't run `--compare=jit` anymore. We
should figure out how to fix that and add `xN` to the fuzzer, even though it's
a fairly new feature it is now documented, see #1572.

I noticed that quickchecks on a function with `xN` was giving odd results,
which was because quickchecks *only* work in the JIT, via IR conversion.
Debugged the discrepancy to `dynamic_cast<BitsType*>` and so knocked out nearly
all of the `dynamic_cast<BitsType*>` and `dynamic_cast<const BitsType*>` in the
code base.  We should also make DSLX concrete type visitor use more pervasive
and consolidate on recommended (vs deprecated, with pointer at what's
recommended) routines as we can.

In that vein this change removes `IsUBits` and `IsSBits` -- less identity
testing and more polymorphic-property-testing the better.